fix: compute the navigable page count in long arithmetic
NavigablePages evaluated (maxOffset / limit) + 1 in unchecked int. A ceiling
written as int.MaxValue -- the spelling a consumer reaches for to lift a shared
PaginateConfigDefaults ceiling for one resource -- makes that expression wrap to
int.MinValue at limit 1, and Math.Min then returns it. meta.hasNextPage went
false on page 1 of 8, links.next went null and links.last pointed back at page 1,
while meta.totalPages still said 8: the clamp contradicted the guard it exists to
agree with, silently and with no error anywhere. ApplyCeiling and ApplyPage
already widen the same class of arithmetic; this is the third site.

Also adds the in-table deep-page case the suite never had. The existing deep-page
test skips 4990 rows out of 8, so the count short-circuit answers it and the
paged query is never issued -- it pins the branch where no offset is applied. The
new case skips 6 of 8, asserts the offset query really runs and returns the right
rows under a ceiling that allows it, and that one row shallower the same request
is refused without the database being asked.

+++ src/Janzen.Pagination.EntityFrameworkCore/PaginateQueryableExtensions.cs
@@ -279,7 +279,10 @@ public static class PaginateQueryableExtensions {
 		// An unlimited read is one page and never skips, so the offset ceiling has nothing to say about it.
 		if (limit == PaginateQuery.UnlimitedLimit || config.MaxOffset is not { } maxOffset) return totalPages;
 
-		return Math.Min(totalPages, (maxOffset / limit) + 1);
+		// Long arithmetic for the same reason ApplyCeiling needs it: at limit 1 a ceiling of int.MaxValue makes
+		// the + 1 wrap to int.MinValue, which then wins the Math.Min -- so a ceiling meaning "no practical cap"
+		// clamped navigation to one page rather than leaving it alone.
+		return (int)Math.Min(totalPages, ((long)maxOffset / limit) + 1);
 
 	}
